fix: move out of internal app state, rename, split
